requirements.txt
Flask==3.1.1
pandas==2.1.4
Pillow==10.1.0
Werkzeug==3.0.1
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
